fix for jitter diagonal was wrong

+++ GameEngine/GameEngine/Elements/Player.cs
@@ -99,8 +99,8 @@ internal class Player : Object
                 direction.Normalize();
             }
             
-            X += (int) (direction.X * Speed * elapsedTime);
-            Y += (int) (direction.Y * Speed * elapsedTime);
+            X += direction.X * Speed * elapsedTime;
+            Y += direction.Y * Speed * elapsedTime;
         }
 
         void UpdateAnimation()
